% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{load_mask_key}
\alias{load_mask_key}
\title{Load a masking key from disk}
\usage{
load_mask_key(path)
}
\arguments{
\item{path}{Path to an \code{.rds} file previously written by \code{\link[=save_mask_key]{save_mask_key()}}.}
}
\value{
A \code{cdiscmask_key} object.
}
\description{
Load a masking key from disk
}
\examples{
key <- new_mask_key()
tmp <- tempfile(fileext = ".rds")
save_mask_key(key, tmp)
key2 <- load_mask_key(tmp)
identical(key$id_map, key2$id_map)
}
