% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{save_mask_key}
\alias{save_mask_key}
\title{Save a masking key to disk}
\usage{
save_mask_key(key, path)
}
\arguments{
\item{key}{A \code{cdiscmask_key} object.}

\item{path}{File path for the \code{.rds} file.}
}
\description{
Save a masking key to disk
}
\examples{
key <- new_mask_key()
tmp <- tempfile(fileext = ".rds")
save_mask_key(key, tmp)
}
